requirements.txt
pandas
numpy
scikit-learn
lightgbm
sqlalchemy
pymysql
mlflow
fastapi
uvicorn
streamlit
shap
python-dotenv
catboost
optuna
loguru
typer
ruff
